perf(vendor): merge vendor bundle with main
This merges the vendor bundle and main bundle together, in order to
make it easier to load all of the initial assets via a single script
tag using the `async` attribute. Using the `async` attribute, helps
webkit unblock the rendering of the page as remote scripts are loaded &
executed. Justification for `async` vs `defer` in following commit 
message.

This does have one slight cache disadvantage, which is that the main 
bundle will likely change at a greater frequency than the vendor bundle,
but now the entire bundle will need to be invalidated with each update.
Given the low frequency at which the app is re-deployed, this isn’t a
high cost.