perf(scripts): make main script load with async attribute
This commit adds the `async` attribute to the only remote script on the
page, so that the browser can continue parsing and rendering the
document without waiting on the script to load and execute. The `async`
attribute is preferable to `defer` because `defer` will block DOM
parsing and rendering in Safari, whereas `async` will not. Since
`async` does not guarantee script execution order like `defer`, scripts
that depend on each other must be bundled and loaded together.