Some loading optimizations (#69)
* perf(sass): move variables into separate file, and only import mixins where needed

Previously, the _themes.scss file included both variables, and @includes, which would cause
redundant CSS rules to be placed into every generated CSS file. 

This commit moves variables declarations into a new file, _theme_variables.scss, which can
safely be included in components or any other SASS file without side effects.

This reduced the total size of all chunks from 830,306 bytes to 484,706 (42% decrease).

* perf(feed): move feed routes into app.module

Since the feed routes are required for the first screen of the 
application, lazy loading the feed components will most likely add more 
latency to the time to bootstrap the home page. This commit moves the
feed components into the AppModule.

* perf(bundles): remove unneeded bundles and imports

 * Remove the environment.dev.ts, which isn’t being used (as far as I 
   could tell).
 * Move dev-specific polyfills (reflect and zone) to dev environment 
   (environment.ts) to reduce size of prod bundle since reflect is only 
   needed in JIT mode.
 * Use minified zone in prod environment (environment.prod.ts)
 * Get rid of now-empty polyfills.ts (and remove its import from 
   main.ts)

* perf(API): use smaller fetch polyfill (unfetch)

* perf(inline): inline the inline so it’s inlined

This inlines webpack’s inline script, which provides the required 
functions for other webpack bundles. The script is so small that it’s 
easier to just include it in the index page.

* perf(vendor): merge vendor bundle with main

This merges the vendor bundle and main bundle together, in order to
make it easier to load all of the initial assets via a single script
tag using the `async` attribute. Using the `async` attribute, helps
webkit unblock the rendering of the page as remote scripts are loaded &
executed. Justification for `async` vs `defer` in following commit 
message.

This does have one slight cache disadvantage, which is that the main 
bundle will likely change at a greater frequency than the vendor bundle,
but now the entire bundle will need to be invalidated with each update.
Given the low frequency at which the app is re-deployed, this isn’t a
high cost.

* perf(scripts): make main script load with async attribute

This commit adds the `async` attribute to the only remote script on the
page, so that the browser can continue parsing and rendering the
document without waiting on the script to load and execute. The `async`
attribute is preferable to `defer` because `defer` will block DOM
parsing and rendering in Safari, whereas `async` will not. Since
`async` does not guarantee script execution order like `defer`, scripts
that depend on each other must be bundled and loaded together.